% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_fill.R
\name{check_and_fill}
\alias{check_and_fill}
\title{check_and_fill function}
\usage{
check_and_fill(data)
}
\arguments{
\item{data}{A newword format data frame.}
}
\description{
This function is to fill the possible conj between the adj and the noun automatically.
}
\details{
Since the function is for fill the word automactically, you need to check oldpro first and give
the class to the cells in evaluation word column which contains '/'.
}
\examples{
data \%<>\% check_and_fill(.)
}
