% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillword.R
\docType{data}
\name{common}
\alias{common}
\title{fillword function}
\format{An object of class \code{data.frame} with 14 rows and 10 columns.}
\usage{
common
}
\arguments{
\item{nchar}{A number. Notice that if nchar = n, you want to put the conj. whose length is less than 2
into the string, you have to key n/2 as input.}

\item{class}{A string. See details.}
}
\description{
This function is for the use of function `check_and_fill`.
}
\keyword{datasets}
