% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{cut}
\alias{cut}
\title{cut function}
\usage{
cut(g)
}
\arguments{
\item{g}{Strings contain ';'}
}
\description{
Cut the synonym in the cells in oldpro and return a string vector contains synonyms.
}
\details{
It will jump error message if there are no ';' in the string.
}
