% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocess.R
\name{dataprocess}
\alias{dataprocess}
\title{data process function}
\usage{
dataprocess(data)
}
\arguments{
\item{data}{data frame.}
}
\description{
This function help you to clean the brackets in old pro and turn the column in data frame
into character vector.
}
\examples{
data \%<>\% dataprocess(.)
}
