% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newword_newpro.R
\name{newword_newpro}
\alias{newword_newpro}
\title{newword_newpro function}
\usage{
newword_newpro(phone, class)
}
\description{
Turn the newword format file into newpro format.
}
\examples{
data_stat = newword_newpro(data, 'stat')
data_key = newword_newpro(data, 'key')
}
