% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_tonewword.R
\name{old_tonewword}
\alias{old_tonewword}
\title{old_tonewword function}
\usage{
old_tonewword(data, where)
}
\arguments{
\item{data}{An oldpro format data frame.}

\item{where}{A string. It will be added as the first column in newword.}
}
\description{
To turn the oldpro into newword format.
}
\details{
Notice that if there are no ';' in any cell in the data frame, this function will return a 0 row data frame without
any error.
}
\examples{
gym = read.csv("...") # gym is an oldpro file
gym_newword = old_tonewword(data = gym, where = "gym")
}
