% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placena.R
\name{placena}
\alias{placena}
\title{placena function}
\usage{
placena(g, nrow)
}
\arguments{
\item{g}{Column contains blank cell.}

\item{nrow}{nrow of the column}
}
\description{
Data frame is able to be written as csv only if there doesn't exist any blank cell in it.
This function is to turn the blank cell into NA value.
}
